refactor(logger.go): separate LoggerWithConfig logic into smaller functions related to #14
