+++ middleware/logger.go
@@ -6,6 +6,7 @@ package middleware
 import (
 	"bytes"
 	"encoding/json"
+	"net/http"
 	"io"
 	"strconv"
 	"strings"
@@ -93,8 +94,8 @@ func Logger() echo.MiddlewareFunc {
 // LoggerWithConfig returns a Logger middleware with config.
 // See: `Logger()`.
 func LoggerWithConfig(config LoggerConfig) echo.MiddlewareFunc {
-	// Defaults
-	if config.Skipper == nil {
+	//Defaults
+	if config.Skipper == nil{
 		config.Skipper = DefaultLoggerConfig.Skipper
 	}
 	if config.Format == "" {
@@ -103,7 +104,6 @@ func LoggerWithConfig(config LoggerConfig) echo.MiddlewareFunc {
 	if config.Output == nil {
 		config.Output = DefaultLoggerConfig.Output
 	}
-
 	config.template = fasttemplate.New(config.Format, "${", "}")
 	config.colorer = color.New()
 	config.colorer.SetOutput(config.Output)
@@ -114,131 +114,161 @@ func LoggerWithConfig(config LoggerConfig) echo.MiddlewareFunc {
 	}
 
 	return func(next echo.HandlerFunc) echo.HandlerFunc {
-		return func(c echo.Context) (err error) {
+		return func(c echo.Context) error {
+			// Se Skipper è vero, salta il middleware
 			if config.Skipper(c) {
 				return next(c)
 			}
-
-			req := c.Request()
-			res := c.Response()
-			start := time.Now()
-			if err = next(c); err != nil {
-				c.Error(err)
-			}
-			stop := time.Now()
-			buf := config.pool.Get().(*bytes.Buffer)
-			buf.Reset()
-			defer config.pool.Put(buf)
-
-			if _, err = config.template.ExecuteFunc(buf, func(w io.Writer, tag string) (int, error) {
-				switch tag {
-				case "custom":
-					if config.CustomTagFunc == nil {
-						return 0, nil
-					}
-					return config.CustomTagFunc(c, buf)
-				case "time_unix":
-					return buf.WriteString(strconv.FormatInt(time.Now().Unix(), 10))
-				case "time_unix_milli":
-					// go 1.17 or later, it supports time#UnixMilli()
-					return buf.WriteString(strconv.FormatInt(time.Now().UnixNano()/1000000, 10))
-				case "time_unix_micro":
-					// go 1.17 or later, it supports time#UnixMicro()
-					return buf.WriteString(strconv.FormatInt(time.Now().UnixNano()/1000, 10))
-				case "time_unix_nano":
-					return buf.WriteString(strconv.FormatInt(time.Now().UnixNano(), 10))
-				case "time_rfc3339":
-					return buf.WriteString(time.Now().Format(time.RFC3339))
-				case "time_rfc3339_nano":
-					return buf.WriteString(time.Now().Format(time.RFC3339Nano))
-				case "time_custom":
-					return buf.WriteString(time.Now().Format(config.CustomTimeFormat))
-				case "id":
-					id := req.Header.Get(echo.HeaderXRequestID)
-					if id == "" {
-						id = res.Header().Get(echo.HeaderXRequestID)
-					}
-					return buf.WriteString(id)
-				case "remote_ip":
-					return buf.WriteString(c.RealIP())
-				case "host":
-					return buf.WriteString(req.Host)
-				case "uri":
-					return buf.WriteString(req.RequestURI)
-				case "method":
-					return buf.WriteString(req.Method)
-				case "path":
-					p := req.URL.Path
-					if p == "" {
-						p = "/"
-					}
-					return buf.WriteString(p)
-				case "route":
-					return buf.WriteString(c.Path())
-				case "protocol":
-					return buf.WriteString(req.Proto)
-				case "referer":
-					return buf.WriteString(req.Referer())
-				case "user_agent":
-					return buf.WriteString(req.UserAgent())
-				case "status":
-					n := res.Status
-					s := config.colorer.Green(n)
-					switch {
-					case n >= 500:
-						s = config.colorer.Red(n)
-					case n >= 400:
-						s = config.colorer.Yellow(n)
-					case n >= 300:
-						s = config.colorer.Cyan(n)
-					}
-					return buf.WriteString(s)
-				case "error":
-					if err != nil {
-						// Error may contain invalid JSON e.g. `"`
-						b, _ := json.Marshal(err.Error())
-						b = b[1 : len(b)-1]
-						return buf.Write(b)
-					}
-				case "latency":
-					l := stop.Sub(start)
-					return buf.WriteString(strconv.FormatInt(int64(l), 10))
-				case "latency_human":
-					return buf.WriteString(stop.Sub(start).String())
-				case "bytes_in":
-					cl := req.Header.Get(echo.HeaderContentLength)
-					if cl == "" {
-						cl = "0"
-					}
-					return buf.WriteString(cl)
-				case "bytes_out":
-					return buf.WriteString(strconv.FormatInt(res.Size, 10))
-				default:
-					switch {
-					case strings.HasPrefix(tag, "header:"):
-						return buf.Write([]byte(c.Request().Header.Get(tag[7:])))
-					case strings.HasPrefix(tag, "query:"):
-						return buf.Write([]byte(c.QueryParam(tag[6:])))
-					case strings.HasPrefix(tag, "form:"):
-						return buf.Write([]byte(c.FormValue(tag[5:])))
-					case strings.HasPrefix(tag, "cookie:"):
-						cookie, err := c.Cookie(tag[7:])
-						if err == nil {
-							return buf.Write([]byte(cookie.Value))
-						}
-					}
-				}
-				return 0, nil
-			}); err != nil {
-				return
-			}
-
-			if config.Output == nil {
-				_, err = c.Logger().Output().Write(buf.Bytes())
-				return
-			}
-			_, err = config.Output.Write(buf.Bytes())
-			return
+			// Processa la richiesta
+			return processRequest(c, next, &config)
 		}
 	}
 }
+
+func processRequest(c echo.Context, next echo.HandlerFunc, config *LoggerConfig) error {
+	req := c.Request()
+	res := c.Response()
+	start := time.Now()
+	err := next(c)
+	if err != nil {
+		c.Error(err)
+	}
+	stop := time.Now()
+
+	// Gestione buffer separata
+	buf := config.pool.Get().(*bytes.Buffer)
+	buf.Reset()
+	defer config.pool.Put(buf)
+
+	// Processa il template
+	if _, err = processTemplate(err, buf, c, req, res, config, start, stop); err != nil {
+		return err
+	}
+
+	// Scrive l'output
+	return writeOutput(buf, c, config)
+}
+
+func processTemplate(err error, buf *bytes.Buffer, c echo.Context, req *http.Request, res *echo.Response, config *LoggerConfig, start, stop time.Time) (int, error) {
+	n, err := config.template.ExecuteFunc(buf, func(w io.Writer, tag string) (int, error) {
+		return handleTag(err, buf, tag, c, req, res, config, start, stop)
+	})
+	return int(n), err
+}
+
+func writeOutput(buf *bytes.Buffer, c echo.Context, config *LoggerConfig) error {
+	if config.Output == nil {
+		_, err := c.Logger().Output().Write(buf.Bytes())
+		return err
+	}
+	_, err := config.Output.Write(buf.Bytes())
+	return err
+}
+
+func handleTag(err error, buf *bytes.Buffer, tag string, c echo.Context, req *http.Request, res *echo.Response, config *LoggerConfig, start, stop time.Time) (int, error) {
+	switch tag {
+	case "custom":
+		if config.CustomTagFunc == nil {
+			return 0, nil
+		}
+		return config.CustomTagFunc(c, buf)
+	case "time_unix":
+		return buf.WriteString(strconv.FormatInt(time.Now().Unix(), 10))
+	case "time_unix_milli":
+		// go 1.17 or later, it supports time#UnixMilli()
+		return buf.WriteString(strconv.FormatInt(time.Now().UnixNano()/1000000, 10))
+	case "time_unix_micro":
+		// go 1.17 or later, it supports time#UnixMicro()
+		return buf.WriteString(strconv.FormatInt(time.Now().UnixNano()/1000, 10))
+	case "time_unix_nano":
+		return buf.WriteString(strconv.FormatInt(time.Now().UnixNano(), 10))
+	case "time_rfc3339":
+		return buf.WriteString(time.Now().Format(time.RFC3339))
+	case "time_rfc3339_nano":
+		return buf.WriteString(time.Now().Format(time.RFC3339Nano))
+	case "time_custom":
+		return buf.WriteString(time.Now().Format(config.CustomTimeFormat))
+	case "id":
+		id := req.Header.Get(echo.HeaderXRequestID)
+		if id == "" {
+			id = res.Header().Get(echo.HeaderXRequestID)
+		}
+		return buf.WriteString(id)
+	case "remote_ip":
+		return buf.WriteString(c.RealIP())
+	case "host":
+		return buf.WriteString(req.Host)
+	case "uri":
+		return buf.WriteString(req.RequestURI)
+	case "method":
+		return buf.WriteString(req.Method)
+	case "path":
+		p := req.URL.Path
+		if p == "" {
+			p = "/"
+		}
+		return buf.WriteString(p)
+	case "route":
+		return buf.WriteString(c.Path())
+	case "protocol":
+		return buf.WriteString(req.Proto)
+	case "referer":
+		return buf.WriteString(req.Referer())
+	case "user_agent":
+		return buf.WriteString(req.UserAgent())
+	case "status":
+		return buf.WriteString(getColoredStatus(res.Status, config))
+	case "error":
+		if err != nil {
+			// Error may contain invalid JSON e.g. `"`
+			b, _ := json.Marshal(err.Error())
+			b = b[1 : len(b)-1]
+			return buf.Write(b)
+		}
+	case "latency":
+		l := stop.Sub(start)
+		return buf.WriteString(strconv.FormatInt(int64(l), 10))
+	case "latency_human":
+		return buf.WriteString(stop.Sub(start).String())
+	case "bytes_in":
+		cl := req.Header.Get(echo.HeaderContentLength)
+		if cl == "" {
+			cl = "0"
+		}
+		return buf.WriteString(cl)
+	case "bytes_out":
+		return buf.WriteString(strconv.FormatInt(res.Size, 10))
+	default:
+		switch {
+		case strings.HasPrefix(tag, "header:"):
+			return buf.Write([]byte(c.Request().Header.Get(tag[7:])))
+		case strings.HasPrefix(tag, "query:"):
+			return buf.Write([]byte(c.QueryParam(tag[6:])))
+		case strings.HasPrefix(tag, "form:"):
+			return buf.Write([]byte(c.FormValue(tag[5:])))
+		case strings.HasPrefix(tag, "cookie:"):
+			cookie, err := c.Cookie(tag[7:])
+			if err == nil {
+				return buf.Write([]byte(cookie.Value))
+			}
+		}	
+	}
+	return 0, nil
+}
+
+
+func getColoredStatus(status int, config *LoggerConfig) string {
+	s := config.colorer.Green(status)
+	switch {
+	case status >= 500:
+		s = config.colorer.Red(status)
+	case status >= 400:
+		s = config.colorer.Yellow(status)
+	case status >= 300:
+		s = config.colorer.Cyan(status)
+	}
+	return s
+}
+
+
+
